(hel-446) Extract algo dans factory
